+++ helloworld-blockchain-core/src/main/java/com/xingkaichun/helloworldblockchain/core/SynchronizerDataBase.java
@@ -2,6 +2,7 @@ package com.xingkaichun.helloworldblockchain.core;
 
 import com.xingkaichun.helloworldblockchain.dto.BlockDTO;
 
+import java.math.BigInteger;
 import java.util.List;
 
 
@@ -53,15 +54,15 @@ public abstract class  SynchronizerDataBase {
     /**
      * 获取节点(nodeId)传输过来所有区块中最小的区块高度。
      */
-    public abstract int getMinBlockHeight(String nodeId) throws Exception ;
+    public abstract BigInteger getMinBlockHeight(String nodeId) throws Exception ;
     /**
      * 获取节点(nodeId)传输过来所有区块中最大的区块高度。
      */
-    public abstract int getMaxBlockHeight(String nodeId) throws Exception ;
+    public abstract BigInteger getMaxBlockHeight(String nodeId) throws Exception ;
     /**
      * 根据节点与区块高度获取区块
      */
-    public abstract BlockDTO getBlockDto(String nodeId,int blockHeight) throws Exception ;
+    public abstract BlockDTO getBlockDto(String nodeId,BigInteger blockHeight) throws Exception ;
     /**
      * 给节点(nodeId)添加数据传输完成的标识。
      */
