+++ helloworld-blockchain-node/src/main/java/com/xingkaichun/helloworldblockchain/node/dto/blockchainbrowser/response/PingResponse.java
@@ -3,11 +3,12 @@ package com.xingkaichun.helloworldblockchain.node.dto.blockchainbrowser.response
 import com.xingkaichun.helloworldblockchain.node.dto.nodeserver.Node;
 import lombok.Data;
 
+import java.math.BigInteger;
 import java.util.List;
 
 @Data
 public class PingResponse {
 
-    private int blockChainHeight ;
+    private BigInteger blockChainHeight ;
     private List<Node> nodeList;
 }
