+++ helloworld-blockchain-node/src/main/java/com/xingkaichun/helloworldblockchain/node/dto/nodeserver/Node.java
@@ -2,10 +2,12 @@ package com.xingkaichun.helloworldblockchain.node.dto.nodeserver;
 
 import lombok.Data;
 
+import java.math.BigInteger;
+
 @Data
 public class Node extends SimpleNode{
 
-    private Integer blockChainHeight;
+    private BigInteger blockChainHeight;
     private Boolean isNodeAvailable;
     private Integer errorConnectionTimes;
     private Boolean fork;
