+++ helloworld-blockchain-node/src/main/java/com/xingkaichun/helloworldblockchain/node/dto/nodeserver/request/QueryBlockDtoByBlockHeightRequest.java
@@ -2,8 +2,10 @@ package com.xingkaichun.helloworldblockchain.node.dto.nodeserver.request;
 
 import lombok.Data;
 
+import java.math.BigInteger;
+
 @Data
 public class QueryBlockDtoByBlockHeightRequest {
 
-    private int blockHeight;
+    private BigInteger blockHeight;
 }
