+++ helloworld-blockchain-node/src/main/java/com/xingkaichun/helloworldblockchain/node/dto/nodeserver/request/QueryBlockHashByBlockHeightRequest.java
@@ -2,9 +2,11 @@ package com.xingkaichun.helloworldblockchain.node.dto.nodeserver.request;
 
 import lombok.Data;
 
+import java.math.BigInteger;
+
 @Data
 public class QueryBlockHashByBlockHeightRequest {
 
-    private int blockHeight;
+    private BigInteger blockHeight;
 
 }
