+++ helloworld-blockchain-node/src/main/java/com/xingkaichun/helloworldblockchain/node/model/NodeEntity.java
@@ -3,10 +3,12 @@ package com.xingkaichun.helloworldblockchain.node.model;
 import com.xingkaichun.helloworldblockchain.node.dto.nodeserver.SimpleNode;
 import lombok.Data;
 
+import java.math.BigInteger;
+
 @Data
 public class NodeEntity extends SimpleNode{
 
-    private Integer blockChainHeight;
+    private BigInteger blockChainHeight;
     private Boolean isNodeAvailable;
     private Integer errorConnectionTimes;
     private Boolean fork;
